Made lots of changes!
Changed the background image to a HQ one, added a backing to the first
paragraph for visibility, tried formatting the layout of the show boxes,
found some real shows and changed photos accordingly, added a hover over
description and link to the website on the first show. But this has
changed the layout of the boxes - I'm working on it!